requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.0
langchain==0.2.5
langchain-community==0.2.5
qdrant-client==1.9.0
redis==5.0.7
email_validator==2.1.1
